Hard code audience to rubygems.org
It was previously https://rubygems.org, and also configurable via
parameters.

When creating an API Key Role on rubygems.org, it defaults the rules to
requiring an `aud` claim with the value `rubygems.org`. By stripping the
`https://` from our default, documenting the process for using this
plugin can have one less step.

It also seems unnecessary to keep in configurable. This plugin is
specifically for rubygems.org integration, and I can't imagine scenarios
where we'd need a different audience on the OIDC tokens.